i386: pc: Move memory map code in its own file
We will also move most generic memory setting code to memory.c.
