+++ include/hw/i386/pc.h
@@ -283,17 +283,6 @@ void pc_system_firmware_init(MemoryRegion *rom_memory,
 void pc_madt_cpu_entry(AcpiDeviceIf *adev, int uid,
                        const CPUArchIdList *apic_ids, GArray *entry);
 
-/* e820 types */
-#define E820_RAM        1
-#define E820_RESERVED   2
-#define E820_ACPI       3
-#define E820_NVS        4
-#define E820_UNUSABLE   5
-
-int e820_add_entry(uint64_t, uint64_t, uint32_t);
-int e820_get_num_entries(void);
-bool e820_get_entry(int, uint32_t, uint64_t *, uint64_t *);
-
 #define PC_COMPAT_2_12 \
     HW_COMPAT_2_12 \
     {\
@@ -313,6 +302,7 @@ bool e820_get_entry(int, uint32_t, uint64_t *, uint64_t *);
         .property = "xlevel",\
         .value    = stringify(0x8000000a),\
     },
+
 
 #define PC_COMPAT_2_11 \
     HW_COMPAT_2_11 \
